Add Delete task into BeforeBuild target on both projects
The following files should be deleted:
1. obj\project.assets.json in noth projects
This file is left after switching from a .NET Standard 2.1 branch  to .NET framework one because it is not  under Git control. This file is left after NET Standard 2.1 and prevents successfull builds for .NET Framework Targets.
2.  $(MSBuildProjectExtensionsPath)$(ProjectFileName).nuget.g.props only in test project. It changes the type of project from Class Library (suitable for Framework)  to Console Application (suitable for Core) and adds aut-generated file to @(Compile) item list, that file cannot be compiled in Framework project.
In this version the first run of Build tries to compile that file though so we need compile test project twice after switching from a Core branch